Add has_feasible() to smp::rrg class
It checks whether there is at least one trajectory satisfying the
mu calculus formula. The return type is bool. The motivation to create
this method is that the only substantial alternative is get_solution()
of the class model_checker_mu_calculus, which will construct a
solution, not merely declare whether there is one.