Improve style and whitespace usage; comment-out status cout
A variety of little notification messages have been commented-out,
i.e., removed as code and placed into a comment block. The motivation
is to allow dumping the solution to stdout, saving it to a file, and
then loading easily from that file. The messages may later be re-
enabled via logging. Also note that the original code included already
cout statements behind comment blocks, so hiding more similarly is
consistent with the existing style.