BUGFIX: Check for redundant state in inter-vertex trajectory
Trajectories between vertices of the sampled graph are defined to be
the states and corresponding inputs to move between the states of the
endpoint vertices. The insert_trajectory() method of the base planner
class did not enforce non-repetition, so that at least for the single
integrator extender, it is possible for trajectory_in to contain
also the state of the destination vertex. This changeset now checks
for this possibility and removes the last state in trajectory_in if
it is equal to vertex_dst_in.

However, this may not be a bug in situations where repetition, i.e.,
staying at the same state for multiple steps, is desired. Such a
special case will need to be treated later.

Note also that inputs are not treated appropriately in this changeset,
but they are not treated correctly or consistently in various other
locations of the code; this is future work.