Correct treatment of inputs for single integrator extender
The documentation of the representation and meaning of input vectors for
the single integrator extender was different from the implementation,
which was basically not useful and incomplete. This changeset makes the
implementation consistent with the documentation. We now make the design
choice to regard the trajectory between the states of adjacent vertices as
requiring unit time for this extender.

If support for bounds on input is later provided, then this assumption of
unit time is easily removed.