dump_json() for saving results from RRG planner
dump_json() prints in JSON format to std::cout the sampled graph entirely in
terms of states and a solution if found by the RRG planner. The intermediate
states corresponding to edges are also included. To facilitate this,
dump_states_json() is introduced as a method of the trajectory class.

N.B., the output format is not completely defined yet and thus in this
commit it is not assigned a version number. The next major step is to
improve the examples, after which we will declare version 0.