Move libbot extension and examples to contrib/
They are broken and, given that they depend on an old version of libbot that
the current version is not backwards-compatible with, it may be a while
before they are made to work again.

Note that various other remaining references to the libbot extension and
examples will be gradually moved here (under contrib/) as the source tree
continues to be reorganized.