EXAMPLE: All but the Dubins aircraft RRG example are operational
including comments indicating when values are chosen to match a hard-coded
assumption elsewhere.

This commit is a bit sloppy in that it mixes style, whitespace, and
nontrivial source code changes into one. But dissecting them in this
particular case is not worth the effort.

Each example outputs results using the format introduced in commit
728a8b3e8149c0740d83d2bb2c50e1d88ab54191