Support for 3D workspaces
as demonstrated by the examples dubins_double_integrator_airplane and
single_integrator_mu_calculus, where the latter can be configured for
either 2- or 3-D by adjusting the value of NUM_DIMENSIONS in
standalone_rrg_single_integrator_mu_calculus.cpp.

NOTE that this changeset breaks operation of the other examples under rrg/
since still the region labeling is hard-coded. Correcting this, i.e.,
making region labeling configurable, is the next major change.

3D plotting in tools/rrgplot.py is achieved using Mayavi
(http://docs.enthought.com/mayavi/mayavi/).

In the output file format, vertices in the sampled graph are now defined as
an object that is the value of the key "V". This simplifies finding and
enumerating the vertices as well as the addition of new top-level keys to
the JSON file.

dump_json() now has a parameter include_graph (bool) by which the user can
choose whether to include the sampled graph in the output. Default is true,
so default behavior is unchanged. The motivation is that the file can be
quite large for the examples like dubins_double_integrator_airplane.

The "obstacles" key of the input format has been renamed to "regions".