Atomic proposition labeling for the case of disjoint regions
The labeling of regions (also called hyper-rectangles or boxes) of the
output space by atomic propositions is now automatic, provided that the
regions are disjoint. To achieve this, I

1. include a region typedef in typeparams, so that collision_checker need
   not be templated with NUM_DIMENIONS; note that this region has number of
   dimensions matching that of the output space (also known as workspace);

2. create a negated atomic proposition type in the incremental model
   checker; the "type" manifests as a node in the formula parse tree;

3. move get_region_index() to be a public member of the class
   collision_checker_mu_calculus;

4. associate an instance of collision_checker_mu_calculus with
   model_checker_mu_calculus, registered via the new public function of the
   latter, add_labeler()

The crucial assumption is made that the dimensions of the state over which
labeling is defined occur first, which is consistent with the existing
examples/rrg/; e.g., for dubins_double_integrator_airplane, the first three
state dimensions are the usual "x,y,z" position, and thus we define
labeling regions to have number of dimensions 3.