Corrected and completed algorithm based on the ACC 2012 paper
There are several issues with the algorithm presented in that paper,
all of which are addressed in this implementation. Besides corrections
of errors, an algorithm for propagating costs following a change in
the Pr data structure is now implemented in ModelCheckerPG::propagateCost().

Getting a winning trajectory for reaching a literal vertex in the game
arena is not yet implemented, and an attempt to do so will cause an
assertion failure.

Demonstrations of usage are under examples/rrgstar/