Correct the formula hard-coded in parseFormulaLoop4()
...and attempt to correct comments, which were not consistent. Before
this changeset, subformula (p1 and z) and (p2 and z) did not require
more transitions after reaching a state labeled with p1 (respectively,
p2). In previous tests this was not practically apparent because p1
and p2 were not among labels of the same states of the Kripke
structure. When they appear in a label for the same state, the
resulting game graph has a loop of vertices all of which have that
same state yet which corresponds to a winning strategy. Hence a
zero-cost suffix is found that corresponds to a single state and thus
is not a permissible trajectory.

This changeset modifies the subformulae to be (p1 and suc z) and
(p2 and suc z).