Find minimum cost trajectory, instead of min loop then prefix
Before this changeset, a trajectory was found by searching for the
lowest cost loop (suffix), and then searching for the lowest cost
prefix that reaches it. Now a minimum cost trajectory is actually
found because comparison is made among entire trajectories (i.e.,
prefix and suffix pairs).

Also, include the prefix in the saved least cost found thus far.