mu-Calculus formula generation for reach-avoid tasks
The template is based on having a collection of regions to repeatedly
visit ("goals") and regions to avoid ("obstacles"). These are numbered
as atomic propositions p_1, ..., p_k, p_{k+1}, ..., p_{N},
where p_1, ..., p_k are to be reached.

The implementation is based on that of parseFormulaLoop4(). The latter
is subsumed by this template, and thus it is deleted.

mu-Calculus model checkers can be instantiated with a number of goals
and number of obstacles. The motivation for supporting this in
addition to having a general formula parser is that it is a common
use-case and simplifies usage. The corresponding mu-Calculus formula
is somewhat long and not obvious.

Default behavior is unchanged.