Correct reach-avoid formula template
This revises the new formula that had been introduced in
a1837a65deeeb9100c8516ebba0e1f7d92d2bb96.

Use only one succ operator and place it immediately after (to the
right of) the LFP operator, once per region to be reached repeatedly.
This enforces continued motion while allowing regions to intersect.

The original implementation of parseFormulaLoop4() should not be
considered broken with respect to the ACC 2012 and CDC 2009 papers
because there the authors described the regions as "distinct subsets."